// packages/astro/dist/types/public/internal.d.ts
export interface RoutePart {
	content: string;
	dynamic: boolean;
	spread: boolean;
}